refactor: 등록모임 페이지를 서버 컴포넌트로 변경
